WIP. Refactor naming of survey and snapshots views and models
