--- src/app/snapshots/add/collection.js
@@ -1,6 +1,0 @@
-import Bn from 'backbone';
-import env from '../../env';
-
-export default Bn.Collection.extend({
-  url: `${env.API}/surveys`
-});
