--- src/app/snapshots/add/new_survey_definition.js
@@ -1,6 +1,0 @@
-import Bn from 'backbone';
-import env from '../../env';
-
-export default Bn.Model.extend({
-  urlRoot: `${env.API}/surveys`
-});
