+++ src/app/snapshots/list/template.hbs
@@ -1,6 +1,6 @@
 <div>
   <div class="text-xs-right pull-right">
-    <button class="btn btn-primary" id="new-survey"> Answer Survey
+    <button class="btn btn-primary" id="new-snapshot"> Answer Survey
     </button>
   </div>
   <br />
